# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.2.0
  description: API specification for RGI Strategic Intelligence System
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /professors/config:
    get:
      operationId: getProfessorLibraryConfig
      tags: [professors]
      summary: Get Professor Library configuration
      security:
        - FirebaseEditorAuth: []
        - InternalServiceKey: []
      responses:
        "200":
          description: Professor Library configuration
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProfessorLibraryConfigResponse"
        "401":
          $ref: "#/components/responses/InternalEditorUnauthorized"
        "403":
          $ref: "#/components/responses/InternalEditorForbidden"
        "503":
          $ref: "#/components/responses/InternalEditorAuthUnavailable"
  /professors:
    get:
      operationId: listProfessorProfiles
      tags: [professors]
      summary: List professor profiles
      security:
        - FirebaseEditorAuth: []
        - InternalServiceKey: []
      parameters:
        - name: status
          in: query
          description: Filter by whether a professor is included in future article matching.
          schema:
            type: string
            enum: [active, inactive]
      responses:
        "200":
          description: Professor profile list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ListProfessorProfilesResponse"
        "401":
          $ref: "#/components/responses/InternalEditorUnauthorized"
        "403":
          $ref: "#/components/responses/InternalEditorForbidden"
        "503":
          $ref: "#/components/responses/InternalEditorAuthUnavailable"
    post:
      operationId: createProfessorProfile
      tags: [professors]
      summary: Create a professor profile
      security:
        - FirebaseEditorAuth: []
        - InternalServiceKey: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProfessorProfileBody"
      responses:
        "201":
          description: Created professor profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProfessorProfileDetailResponse"
        "401":
          $ref: "#/components/responses/InternalEditorUnauthorized"
        "403":
          $ref: "#/components/responses/InternalEditorForbidden"
        "503":
          $ref: "#/components/responses/InternalEditorAuthUnavailable"
  /professors/{id}:
    get:
      operationId: getProfessorProfile
      tags: [professors]
      summary: Get a professor profile
      security:
        - FirebaseEditorAuth: []
        - InternalServiceKey: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            pattern: '^[A-Za-z0-9_-]{8,128}$'
      responses:
        "200":
          description: Professor profile detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProfessorProfileDetailResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          $ref: "#/components/responses/InternalEditorUnauthorized"
        "403":
          $ref: "#/components/responses/InternalEditorForbidden"
        "503":
          $ref: "#/components/responses/InternalEditorAuthUnavailable"
    patch:
      operationId: updateProfessorProfile
      tags: [professors]
      summary: Update a professor profile
      security:
        - FirebaseEditorAuth: []
        - InternalServiceKey: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            pattern: '^[A-Za-z0-9_-]{8,128}$'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProfessorProfileBody"
      responses:
        "200":
          description: Updated professor profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProfessorProfileDetailResponse"
        "401":
          $ref: "#/components/responses/InternalEditorUnauthorized"
        "403":
          $ref: "#/components/responses/InternalEditorForbidden"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "503":
          $ref: "#/components/responses/InternalEditorAuthUnavailable"
...
components:
  securitySchemes:
    FirebaseEditorAuth:
      type: http
      scheme: bearer
      bearerFormat: Firebase ID token
      description: Firebase ID token for a user whose UID is present in the server-only approved RGI editor allowlist. Missing verification configuration fails closed.
    InternalServiceKey:
      type: apiKey
      in: header
      name: x-admin-api-key
      description: Trusted server-to-server operational credential. This key must never be embedded in or sent by the browser application.
  responses:
    InternalEditorUnauthorized:
      description: Missing, malformed, invalid, expired, or revoked Firebase ID token, or invalid internal service credential
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    InternalEditorForbidden:
      description: Authenticated user is not an approved RGI editor, or the requested feature is disabled
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    InternalEditorAuthUnavailable:
      description: Server-side Firebase token verification or approved-editor configuration is unavailable
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        code:
          type: string
        retryable:
          type: boolean
        userMessage:
          type: string
      required:
        - error
...
    ProfessorProfile:
      type: object
      properties:
        id:
          type: string
        fullName:
          type: string
          minLength: 1
          maxLength: 160
        academicTitle:
          type: string
          minLength: 1
          maxLength: 160
        department:
          type: string
          minLength: 1
          maxLength: 160
        coursesTaught:
          type: array
          maxItems: 40
          items:
            type: string
            maxLength: 120
        expertiseTags:
          type: array
          maxItems: 40
          items:
            type: string
            maxLength: 80
        researchInterests:
          type: array
          maxItems: 40
          items:
            type: string
            maxLength: 160
        professionalExperienceTags:
          type: array
          maxItems: 40
          items:
            type: string
            maxLength: 120
        academicExperienceTags:
          type: array
          maxItems: 40
          items:
            type: string
            maxLength: 120
        industries:
          type: array
          maxItems: 30
          items:
            type: string
            maxLength: 80
        topicInterests:
          type: array
          maxItems: 40
          items:
            type: string
            maxLength: 120
        regions:
          type: array
          maxItems: 30
          items:
            type: string
            maxLength: 80
        affiliations:
          type: array
          maxItems: 40
          items:
            type: string
            maxLength: 160
        professionalBackground:
          type: string
          maxLength: 2000
        approvedBio:
          type: string
          maxLength: 2000
        publications:
          type: array
          maxItems: 40
          items:
            type: string
            maxLength: 240
        publicationTopicTags:
          type: array
          maxItems: 40
          items:
            type: string
            maxLength: 120
        recurringThemes:
          type: array
          maxItems: 40
          items:
            type: string
            maxLength: 120
        contactableTopics:
          type: array
          maxItems: 40
          items:
            type: string
            maxLength: 120
        restrictedTopics:
          type: array
          maxItems: 40
          items:
            type: string
            maxLength: 120
        doNotContactTopics:
          type: array
          maxItems: 40
          items:
            type: string
            maxLength: 120
        institutionalConflicts:
          type: array
          maxItems: 40
          items:
            type: string
            maxLength: 160
        affiliationConcerns:
          type: array
          maxItems: 40
          items:
            type: string
            maxLength: 160
        status:
          type: string
          description: Active profiles are included in future article matching; inactive profiles are retained but excluded.
          enum: [active, inactive]
        schemaVersion:
          type: integer
          minimum: 1
        profileRevision:
          type: integer
          minimum: 1
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - fullName
        - academicTitle
        - department
        - coursesTaught
        - expertiseTags
        - researchInterests
        - professionalExperienceTags
        - academicExperienceTags
        - industries
        - topicInterests
        - regions
        - affiliations
        - professionalBackground
        - approvedBio
        - publications
        - publicationTopicTags
        - recurringThemes
        - contactableTopics
        - restrictedTopics
        - doNotContactTopics
        - institutionalConflicts
        - affiliationConcerns
        - status
        - schemaVersion
        - profileRevision
        - createdAt
        - updatedAt
    ProfessorLibraryConfigResponse:
      type: object
      properties:
        writesEnabled:
          type: boolean
      required:
        - writesEnabled
    ListProfessorProfilesResponse:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/ProfessorProfile"
        total:
          type: integer
        writesEnabled:
          type: boolean
      required:
        - items
        - total
        - writesEnabled
    ProfessorProfileDetailResponse:
      type: object
      properties:
        profile:
          $ref: "#/components/schemas/ProfessorProfile"
        writesEnabled:
          type: boolean
      required:
        - profile
        - writesEnabled
    CreateProfessorProfileBody:
      type: object
      additionalProperties: false
      properties:
        fullName:
          type: string
          minLength: 1
          maxLength: 160
        academicTitle:
          type: string
          minLength: 1
          maxLength: 160
        department:
          type: string
          minLength: 1
          maxLength: 160
        coursesTaught:
          type: array
          items:
            type: string
        expertiseTags:
          type: array
          items:
            type: string
        researchInterests:
          type: array
          items:
            type: string
        professionalExperienceTags:
          type: array
          items:
            type: string
        academicExperienceTags:
          type: array
          items:
            type: string
        industries:
          type: array
          items:
            type: string
        topicInterests:
          type: array
          items:
            type: string
        regions:
          type: array
          items:
            type: string
        affiliations:
          type: array
          items:
            type: string
        professionalBackground:
          type: string
        approvedBio:
          type: string
        publications:
          type: array
          items:
            type: string
        publicationTopicTags:
          type: array
          items:
            type: string
        recurringThemes:
          type: array
          items:
            type: string
        contactableTopics:
          type: array
          items:
            type: string
        restrictedTopics:
          type: array
          items:
            type: string
        doNotContactTopics:
          type: array
          items:
            type: string
        institutionalConflicts:
          type: array
          items:
            type: string
        affiliationConcerns:
          type: array
          items:
            type: string
        status:
          type: string
          description: Active profiles are included in future article matching; inactive profiles are retained but excluded.
          enum: [active, inactive]
      required:
        - fullName
        - academicTitle
        - department
    UpdateProfessorProfileBody:
      type: object
      additionalProperties: false
      minProperties: 1
      properties:
        fullName:
          type: string
          minLength: 1
          maxLength: 160
        academicTitle:
          type: string
          minLength: 1
          maxLength: 160
        department:
          type: string
          minLength: 1
          maxLength: 160
        coursesTaught:
          type: array
          items:
            type: string
        expertiseTags:
          type: array
          items:
            type: string
        researchInterests:
          type: array
          items:
            type: string
        professionalExperienceTags:
          type: array
          items:
            type: string
        academicExperienceTags:
          type: array
          items:
            type: string
        industries:
          type: array
          items:
            type: string
        topicInterests:
          type: array
          items:
            type: string
        regions:
          type: array
          items:
            type: string
        affiliations:
          type: array
          items:
            type: string
        professionalBackground:
          type: string
        approvedBio:
          type: string
        publications:
          type: array
          items:
            type: string
        publicationTopicTags:
          type: array
          items:
            type: string
        recurringThemes:
          type: array
          items:
            type: string
        contactableTopics:
          type: array
          items:
            type: string
        restrictedTopics:
          type: array
          items:
            type: string
        doNotContactTopics:
          type: array
          items:
            type: string
        institutionalConflicts:
          type: array
          items:
            type: string
        affiliationConcerns:
          type: array
          items:
            type: string
        status:
          type: string
          description: Active profiles are included in future article matching; inactive profiles are retained but excluded.
          enum: [active, inactive]
...